gdal CLI: remove static registration of top-level algorithms
This should fix https://github.com/firelab/gdalraster/issues/826

Reading https://stackoverflow.com/questions/32584406/static-initializer-is-optimized-away-when-it-is-in-a-library
and https://stackoverflow.com/questions/1804606/static-initialization-and-destruction-of-a-static-librarys-globals-not-happenin
I've come to the conclusion we used the pattern of the later, which
yields to issue when GDAL is built as a static library and used by other
code, with the compilation units where the static registration was done
being discarded by the linker.
