+++ apps/gdalalg_dataset.cpp
@@ -10,7 +10,9 @@
  * SPDX-License-Identifier: MIT
  ****************************************************************************/
 
-#include "gdalalgorithm.h"
+//! @cond Doxygen_Suppress
+
+#include "gdalalg_dataset.h"
 
 #include "gdalalg_dataset_identify.h"
 #include "gdalalg_dataset_copy.h"
@@ -21,33 +23,15 @@
 /*                         GDALDatasetAlgorithm                         */
 /************************************************************************/
 
-class GDALDatasetAlgorithm final : public GDALAlgorithm
+GDALDatasetAlgorithm::GDALDatasetAlgorithm()
+    : GDALAlgorithm(NAME, DESCRIPTION, HELP_URL)
 {
-  public:
-    static constexpr const char *NAME = "dataset";
-    static constexpr const char *DESCRIPTION = "Commands to manage datasets.";
-    static constexpr const char *HELP_URL = "/programs/gdal_dataset.html";
-
-    GDALDatasetAlgorithm() : GDALAlgorithm(NAME, DESCRIPTION, HELP_URL)
-    {
-        RegisterSubAlgorithm<GDALDatasetIdentifyAlgorithm>();
-        RegisterSubAlgorithm<GDALDatasetCopyAlgorithm>();
-        RegisterSubAlgorithm<GDALDatasetRenameAlgorithm>();
-        RegisterSubAlgorithm<GDALDatasetDeleteAlgorithm>();
-    }
-
-    ~GDALDatasetAlgorithm() override;
-
-  private:
-    bool RunImpl(GDALProgressFunc, void *) override
-    {
-        CPLError(CE_Failure, CPLE_AppDefined,
-                 "The Run() method should not be called directly on the \"gdal "
-                 "dataset\" program.");
-        return false;
-    }
-};
+    RegisterSubAlgorithm<GDALDatasetIdentifyAlgorithm>();
+    RegisterSubAlgorithm<GDALDatasetCopyAlgorithm>();
+    RegisterSubAlgorithm<GDALDatasetRenameAlgorithm>();
+    RegisterSubAlgorithm<GDALDatasetDeleteAlgorithm>();
+}
 
 GDALDatasetAlgorithm::~GDALDatasetAlgorithm() = default;
 
-GDAL_STATIC_REGISTER_ALG(GDALDatasetAlgorithm);
+//! @endcond
