+++ apps/gdalalg_raster.cpp
@@ -10,7 +10,9 @@
  * SPDX-License-Identifier: MIT
  ****************************************************************************/
 
-#include "gdalalgorithm.h"
+//! @cond Doxygen_Suppress
+
+#include "gdalalg_raster.h"
 
 #include "gdalalg_raster_info.h"
 #include "gdalalg_raster_aspect.h"
@@ -67,72 +69,59 @@
 /*                         GDALRasterAlgorithm                          */
 /************************************************************************/
 
-class GDALRasterAlgorithm final : public GDALAlgorithm
+GDALRasterAlgorithm::GDALRasterAlgorithm()
+    : GDALAlgorithm(NAME, DESCRIPTION, HELP_URL)
 {
-  public:
-    static constexpr const char *NAME = "raster";
-    static constexpr const char *DESCRIPTION = "Raster commands.";
-    static constexpr const char *HELP_URL = "/programs/gdal_raster.html";
+    AddArg("drivers", 0, _("Display raster driver list as JSON document"),
+           &m_drivers);
 
-    GDALRasterAlgorithm() : GDALAlgorithm(NAME, DESCRIPTION, HELP_URL)
-    {
-        AddArg("drivers", 0, _("Display raster driver list as JSON document"),
-               &m_drivers);
+    AddOutputStringArg(&m_output);
 
-        AddOutputStringArg(&m_output);
-
-        RegisterSubAlgorithm<GDALRasterInfoAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterAsFeaturesAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterAspectAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterBlendAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterCalcAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterCleanCollarAlgorithm>();
-        RegisterSubAlgorithm<GDALRasterClipAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterColorMapAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterCompareAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterConvertAlgorithm>();
-        RegisterSubAlgorithm<GDALRasterCreateAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterEditAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterFootprintAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterHillshadeAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterFillNodataAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterIndexAlgorithm>();
-        RegisterSubAlgorithm<GDALRasterNeighborsAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterOverviewAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterPipelineAlgorithm>();
-        RegisterSubAlgorithm<GDALRasterPixelInfoAlgorithm>();
-        RegisterSubAlgorithm<GDALRasterProximityAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterRGBToPaletteAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterReclassifyAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterReprojectAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterMosaicAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterNoDataToAlphaAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterPansharpenAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterPolygonizeAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterResizeAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterRoughnessAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterContourAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterScaleAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterSelectAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterSetTypeAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterSieveAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterSlopeAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterStackAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterTileAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterTPIAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterTRIAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterUnscaleAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterUpdateAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterViewshedAlgorithmStandalone>();
-        RegisterSubAlgorithm<GDALRasterZonalStatsAlgorithmStandalone>();
-    }
-
-  private:
-    std::string m_output{};
-    bool m_drivers = false;
-
-    bool RunImpl(GDALProgressFunc, void *) override;
-};
+    RegisterSubAlgorithm<GDALRasterInfoAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterAsFeaturesAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterAspectAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterBlendAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterCalcAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterCleanCollarAlgorithm>();
+    RegisterSubAlgorithm<GDALRasterClipAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterColorMapAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterCompareAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterConvertAlgorithm>();
+    RegisterSubAlgorithm<GDALRasterCreateAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterEditAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterFootprintAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterHillshadeAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterFillNodataAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterIndexAlgorithm>();
+    RegisterSubAlgorithm<GDALRasterNeighborsAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterOverviewAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterPipelineAlgorithm>();
+    RegisterSubAlgorithm<GDALRasterPixelInfoAlgorithm>();
+    RegisterSubAlgorithm<GDALRasterProximityAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterRGBToPaletteAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterReclassifyAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterReprojectAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterMosaicAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterNoDataToAlphaAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterPansharpenAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterPolygonizeAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterResizeAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterRoughnessAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterContourAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterScaleAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterSelectAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterSetTypeAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterSieveAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterSlopeAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterStackAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterTileAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterTPIAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterTRIAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterUnscaleAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterUpdateAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterViewshedAlgorithmStandalone>();
+    RegisterSubAlgorithm<GDALRasterZonalStatsAlgorithmStandalone>();
+}
 
 bool GDALRasterAlgorithm::RunImpl(GDALProgressFunc, void *)
 {
@@ -150,4 +139,4 @@ bool GDALRasterAlgorithm::RunImpl(GDALProgressFunc, void *)
     }
 }
 
-GDAL_STATIC_REGISTER_ALG(GDALRasterAlgorithm);
+//! @endcond
