+++ apps/gdalalg_raster.h
@@ -1,0 +1,42 @@
+/******************************************************************************
+ *
+ * Project:  GDAL
+ * Purpose:  gdal "raster" subcommand
+ * Author:   Even Rouault <even dot rouault at spatialys.com>
+ *
+ ******************************************************************************
+ * Copyright (c) 2024, Even Rouault <even dot rouault at spatialys.com>
+ *
+ * SPDX-License-Identifier: MIT
+ ****************************************************************************/
+
+#include "gdalalgorithm.h"
+
+#ifndef GDALALG_RASTER_INCLUDED
+#define GDALALG_RASTER_INCLUDED
+
+//! @cond Doxygen_Suppress
+
+/************************************************************************/
+/*                         GDALRasterAlgorithm                          */
+/************************************************************************/
+
+class GDALRasterAlgorithm final : public GDALAlgorithm
+{
+  public:
+    static constexpr const char *NAME = "raster";
+    static constexpr const char *DESCRIPTION = "Raster commands.";
+    static constexpr const char *HELP_URL = "/programs/gdal_raster.html";
+
+    GDALRasterAlgorithm();
+
+  private:
+    std::string m_output{};
+    bool m_drivers = false;
+
+    bool RunImpl(GDALProgressFunc, void *) override;
+};
+
+//! @endcond
+
+#endif
